requirements.txt
django==3.2.12
django-widget-tweaks==1.4.8
cvss==2.3
django-extra-views==0.14.0
djangorestframework==3.12.4
python-libnmap==0.7.2
Pillow==9.0.0
markdown==3.3.6
bleach==4.1.0
gunicorn==20.1.0
markdownify==0.10.0
weasyprint==53.4
django-crispy-forms==1.13.0
crispy-bootstrap5
django-guardian==2.4.0
django-taggit==2.0.0
django-split-settings==1.1.0
celery
<<<<<<< HEAD
redis
matplotlib==3.5.1
defusedxml==0.7.1
=======
django-celery-results==2.2.0
redis
matplotlib==3.5.1
defusedxml==0.7.1
django-cors-headers==3.11.0
pygments==2.11.2
django-filter==21.1
>>>>>>> origin/dev
